modif detailmapelcreate dan kelas

+++ app/Views/datamaster/kelas.php
@@ -6,21 +6,23 @@ if (!isset($_SESSION["login"])) {
     exit;
 }
 ?>
-<a href="<?= base_url(); ?>/kelas/create" class="btn btn-primary mt-3 mb-3">Tambah Data Kelas</a>
-<?php if (isset($delete)) : ?>
-    <div class="alert alert-success" role="alert">
-        Data Berhasil Di Hapus
-    </div>
-<?php endif; ?>
-<?php if (isset($edit)) : ?>
-    <div class="alert alert-success" role="alert">
-        Data Berhasil Di Edit
-    </div>
-<?php endif; ?>
-<?php if (isset($tambah)) : ?>
-    <div class="alert alert-success" role="alert">
-        Data Berhasil Di Tambah
-    </div>
+<?php if ($_SESSION["status"] === 'admin') : ?>
+    <a href="<?= base_url(); ?>/kelas/create" class="btn btn-primary mt-3 mb-3">Tambah Data Kelas</a>
+    <?php if (isset($delete)) : ?>
+        <div class="alert alert-success" role="alert">
+            Data Berhasil Di Hapus
+        </div>
+    <?php endif; ?>
+    <?php if (isset($edit)) : ?>
+        <div class="alert alert-success" role="alert">
+            Data Berhasil Di Edit
+        </div>
+    <?php endif; ?>
+    <?php if (isset($tambah)) : ?>
+        <div class="alert alert-success" role="alert">
+            Data Berhasil Di Tambah
+        </div>
+    <?php endif; ?>
 <?php endif; ?>
 <table class="table">
     <thead>
@@ -39,9 +41,11 @@ if (!isset($_SESSION["login"])) {
                 <td><?= $k['nama_kelas'] ?></td>
                 <td><?= $k['nip'] ?></td>
                 <td>
-                    <a href="<?= base_url(); ?>/kelas/edit/<?= $k['id_kelas']; ?>" class="btn btn-warning">Edit</a>
-                    <a href="<?= base_url(); ?>/kelas/delete/<?= $k['id_kelas']; ?>" class="btn btn-danger" onclick="return confirm('apakah anda yakin?');">Delete</a>
-                    <br>
+                    <?php if ($_SESSION["status"] === 'admin') : ?>
+                        <a href="<?= base_url(); ?>/kelas/edit/<?= $k['id_kelas']; ?>" class="btn btn-warning">Edit</a>
+                        <a href="<?= base_url(); ?>/kelas/delete/<?= $k['id_kelas']; ?>" class="btn btn-danger" onclick="return confirm('apakah anda yakin?');">Delete</a>
+                        <br>
+                    <?php endif; ?>
                     <a href="<?= base_url(); ?>/kelasSiswa/index/<?= $k['id_kelas']; ?>/<?= $k['nama_kelas'] ?>" class="btn btn-info mt-2">Data Siswa</a>
                     <a href="<?= base_url(); ?>/kelasMapel/index/<?= $k['id_kelas']; ?>/<?= $k['nama_kelas'] ?>" class="btn btn-info mt-2">Mata Pelajaran</a>
                     <a href="<?= base_url(); ?>/kelasJamPelajaran/index/<?= $k['id_kelas']; ?>/<?= $k['nama_kelas'] ?>" class="btn btn-info mt-2">Jam Pelajaran</a>
